moved stack tests to correct file
